Always use a temp file when posting data
There was an odd hang when posting by sending the data directly to the
process that did not manifest when writing the data to a file and
sending it through that way.  The code is a bit more involved,
especially with cleanup of the written file, but now curl behaves as
expected.